+++ lib/appstate.py
@@ -99,6 +99,9 @@ art_cache_dir = None
 song_pack_art_exists = None
 art_override_paths = None
 art_safe_name = None
+# The canonical settings-defaults builder — stays in server.py (shared with the
+# scan/artist-links code) but the settings router calls it through the seam.
+default_settings = None
 
 _SLOTS = frozenset({
     "meta_db", "audio_effect_mappings", "tuning_providers",
@@ -107,6 +110,7 @@ _SLOTS = frozenset({
     "get_progression_content", "builtin_diagnostic_filename",
     "running_version",
     "art_cache_dir", "song_pack_art_exists", "art_override_paths", "art_safe_name",
+    "default_settings",
 })
